qqww

+++ CS272-H01/tests/checkresults/junit/RoachPopulationTest.java
@@ -10,21 +10,21 @@ public class RoachPopulationTest extends AbstractClassDefinitionTest
 {getTests(getClass());methodData=new String[][]{{"void","spray"},{"void",
 "breed"},{"int","getRoaches"}};fieldData=new String[][] {{"population","int"},
 };constructorData=new String[][]{{"int"},};} public void _test_1(){obj=create(
-thisClass,10);assertEquals(10,invoke(obj,int.class,"getRoaches",noArgs));
+thisClass,10);assertEquals(10,(int)invoke(obj,int.class,"getRoaches",noArgs));
 out("Correctness: constructor initializes population correctly.");}public void 
 _test_2(){obj=create(thisClass,10);invoke(obj,"breed",noArgs);invoke(obj,
-"spray",noArgs);assertEquals(18,invoke(obj,int.class,"getRoaches",noArgs));
+"spray",noArgs);assertEquals(18,(int)invoke(obj,int.class,"getRoaches",noArgs));
 out("Correctness: round #1 simulation correct.");}public void _test_3(){obj=
 create(thisClass,10);invoke(obj,"breed",noArgs);invoke(obj,"spray",noArgs);
-invoke(obj,"breed",noArgs);invoke(obj,"spray",noArgs);assertEquals(33,invoke(
+invoke(obj,"breed",noArgs);invoke(obj,"spray",noArgs);assertEquals(33,(int)invoke(
 obj,int.class,"getRoaches",noArgs));out("Correctness: round #2 simulation " + 
 "correct.");} public void _test_4() {obj=create(thisClass,10);invoke(obj,
 "breed",noArgs);invoke(obj,"spray",noArgs);invoke(obj,"breed",noArgs);invoke(
 obj,"spray",noArgs);invoke(obj,"breed",noArgs);invoke(obj,"spray",noArgs);
-assertEquals(60,invoke(obj,int.class,"getRoaches",noArgs));out("Correctness:" + 
+assertEquals(60,(int)invoke(obj,int.class,"getRoaches",noArgs));out("Correctness:" + 
 " round #3 simulation correct.");} public void _test_5(){obj=create(thisClass,
 10);invoke(obj,"breed",noArgs);invoke(obj,"spray",noArgs);invoke(obj,"breed",
 noArgs);invoke(obj,"spray",noArgs);invoke(obj,"breed",noArgs);invoke(obj,
 "spray",noArgs);invoke(obj,"breed",noArgs);invoke(obj,"spray",noArgs);
-assertEquals(108,invoke(obj,int.class,"getRoaches",noArgs));out("Correctness:" +
+assertEquals(108,(int)invoke(obj,int.class,"getRoaches",noArgs));out("Correctness:" +
 " round #4 simulation correct.");}}
